Change: Extract the withFolding HOC into an own module
Keep things separated. With using an own module the HOC code will be
easier to find.
